+++ packages/media-query-list-parser/test/cases/various/0009.mjs
@@ -2,12 +2,13 @@ import assert from 'assert';
 import { runTest } from '../../util/run-test.mjs';
 
 runTest(
-	'(resolution < 1dpi) and (2dpi <= resolution) ',
-	'various/0003',
+	'(resolution < infinite) and (infinite <= resolution) ',
+	'various/0009',
 	(actual, expected) => {
 		assert.deepStrictEqual(
 			actual,
 			expected,
 		);
 	},
+	2,
 );